requirements.txt
# Database
mysql-connector-python>=8.0.0

# Vector Database
qdrant-client>=1.7.0

# Google Gemini AI
google-generativeai>=0.3.0

# Streamlit (for web interface)
streamlit>=1.28.0
pandas>=2.0.0

# Additional dependencies
python-dotenv>=1.0.0
